requirements.txt
streamlit
groq
python-dotenv
pypdf
sentence-transformers
scikit-learn
pandas
matplotlib
reportlab
